+++ src/main/java/com/telepathicgrunt/the_bumblezone/items/CarpenterBeeBoots.java
@@ -62,7 +62,7 @@ public class CarpenterBeeBoots extends BeeArmor {
                 xInBlock < 0.8d &&
                 zInBlock > 0.2d &&
                 zInBlock < 0.8d &&
-                GeneralUtils.isPermissionAllowedAtSpot(world, player, player.blockPosition()))
+                GeneralUtils.isPermissionAllowedAtSpot(world, player, player.blockPosition().below(), false))
             {
                 BlockPos belowBlockPos = new BlockPos(player.position().add(0, -0.1d, 0));
                 BlockState belowBlockState = world.getBlockState(belowBlockPos);