+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/utils/PlatformHooks.java
@@ -127,4 +127,9 @@ public class PlatformHooks {
     public static Player getCraftingPlayer() {
         throw new NotImplementedException("PlatformHooks getCraftingPlayer is not implemented!");
     }
+
+    @ExpectPlatform
+    public static boolean isPermissionAllowedAtSpot(Level level, Entity entity, BlockPos pos, boolean placingBlock) {
+        throw new NotImplementedException("PlatformHooks isPermissionAllowedAtSpot is not implemented!");
+    }
 }