+++ fabric/src/main/java/com/telepathicgrunt/the_bumblezone/utils/fabric/PlatformHooksImpl.java
@@ -200,4 +200,16 @@ public class PlatformHooksImpl {
     public static Player getCraftingPlayer() {
         return PlatformSharedData.craftingPlayer;
     }
+
+    public static boolean isPermissionAllowedAtSpot(Level level, Entity entity, BlockPos pos, boolean placingBlock) {
+        if (placingBlock) {
+            if (entity instanceof Player player) {
+                return PlayerBlockBreakEvents.BEFORE.invoker().beforeBlockBreak(level, player, pos, level.getBlockState(pos), null);
+            }
+        }
+        else if (entity instanceof Player player) {
+            return PlayerBlockBreakEvents.BEFORE.invoker().beforeBlockBreak(level, player, pos, level.getBlockState(pos), null);
+        }
+        return true;
+    }
 }