+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 ### **(V.6.7.6 Changes) (1.19.2 Minecraft)**
 
+##### Items:
+Fixed Carpenter Bee Boots able to mine in protected server spawn chunks by non-op players. Should not happen now.
+
+Fixed Pollen Puff thrown by non-op players able to place Pile of Pollen blocks in protected server spawn chunks. Should not happen now.
+
 ##### Mod Compat:
 Adjusted certain mod's flowers from spawning or not in the Bumblezone's Hanging Gardens structure.
 
 Adjusted some Bee Queen trades to add some more trades using some mod's items.
+
+Fixed Carpenter Bee Boots able to mine in some claimed chunks by non-claim owners players. Only works for some claim mods.
+
+Fixed Pollen Puff thrown by non-claim owner players able to place Pile of Pollen blocks in some claimed chunks. Only works for some claim mods.
 
 
 ### **(V.6.7.5 Changes) (1.19.2 Minecraft)**