+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### **(V.6.7.6 Changes) (1.19.4 Minecraft)**
 
+##### Blocks:
+Fixed String Curtains able to be extended by non-op players in protected server spawn chunks. Should not happen now.
+
 ##### Items:
 (Fabric/Quilt): Fixed crash with Pollen Puff flower spawning. Had to update Fake Player API.
 
@@ -16,9 +19,11 @@ Adjusted certain mod's flowers from spawning or not in the Bumblezone's Hanging 
 
 Adjusted some Bee Queen trades to add some more trades using some mod's items.
 
-Fixed Carpenter Bee Boots able to mine in some claimed chunks by non-claim owners players. Only works for some claim mods.
+Fixed Carpenter Bee Boots able to mine in claimed chunks by non-claim owners players. Only works for some claim mods.
 
-Fixed Pollen Puff thrown by non-claim owner players able to place Pile of Pollen blocks in some claimed chunks. Only works for some claim mods.
+Fixed Pollen Puff thrown by non-claim owner players able to place Pile of Pollen blocks in claimed chunks. Only works for some claim mods.
+
+Fixed String Curtains able to be extended by non-claim owner players in claimed chunks. Only works for some claim mods.
 
 
 ### **(V.6.7.5 Changes) (1.19.4 Minecraft)**